copilot-outputs/requirements.txt
# RAG Dashboard Requirements
streamlit>=1.28.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
openai>=1.3.0
requests>=2.31.0